requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
PyPDF2==3.0.1
python-docx==0.8.11
langchain==0.0.329
chromadb==0.4.15
python-dotenv==1.0.0
openai==0.28.1
torch==2.1.0
protobuf>=5.26.1
packaging<24.0
transformers>=4.31.0
sentence-transformers>=2.2.2
pytesseract==0.3.10
pdf2image==1.16.3
pillow==10.0.0
aiofiles==23.2.1
asyncio==3.4.3